+++ Example/Emission/ARAdminPreloadTableViewCell.h
@@ -1,5 +1,6 @@
 #import <UIKit/UIKit.h>
+#import "ARAdminTableViewCell.h"
 
-@interface ARAdminPreloadTableViewCell : UITableViewCell
+@interface ARAdminPreloadTableViewCell : ARAdminTableViewCell
 @property (nonatomic, copy) dispatch_block_t preloadBlock;
 @end