+++ Example/Emission/ARLabOptions.h
@@ -1,6 +1,6 @@
 #import <Foundation/Foundation.h>
 
-extern NSString *const AROptionsAnExampleLabOption;
+extern NSDictionary *AROptionsAnExampleLabOption;
 
 @interface ARLabOptions : NSObject
 
@@ -8,6 +8,10 @@ extern NSString *const AROptionsAnExampleLabOption;
 + (NSArray *)labsOptions;
 /// A subset of the above that the app should restart upon changing
 + (NSArray *)labsOptionsThatRequireRestart;
+/// A dictionary of lab options to true/false as NSNumbers
++ (NSDictionary *)labOptionsMap;
+/// So we can provide a UI
++ (NSString *)descriptionForOption:(NSString *)option;
 
 /// Get and set individual options
 + (BOOL)boolForOption:(NSString *)option;