+++ Example/Emission/AppDelegate.h
@@ -7,5 +7,7 @@
 // TODO: abstract into a switchboard?
 - (UIViewController *)viewControllerForRoute:(NSString *)route;
 
+// Reloads Emission, and re-creates native modules etc
+- (void)reloadEmission;
 @end